requirements.txt
asgiref==3.9.1
certifi==2025.8.3
charset-normalizer==3.4.3
colorama==0.4.6
defusedxml==0.7.1
Django==5.2.5
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
gunicorn==23.0.0
httpie==3.2.4
idna==3.10
markdown-it-py==4.0.0
mdurl==0.1.2
multidict==6.6.4
packaging==25.0
pillow==11.3.0
psycopg2-binary==2.9.10
Pygments==2.19.2
PyJWT==2.10.1
PySocks==1.7.1
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.1.0
setuptools==80.9.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
